# backend/api/gen/go/admin/service/v1/i_internal_message_category_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_internal_message_category.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-admin/api/gen/go/internal_message/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationInternalMessageCategoryServiceCreate = "/admin.service.v1.InternalMessageCategoryService/Create"
const OperationInternalMessageCategoryServiceDelete = "/admin.service.v1.InternalMessageCategoryService/Delete"
const OperationInternalMessageCategoryServiceGet = "/admin.service.v1.InternalMessageCategoryService/Get"
const OperationInternalMessageCategoryServiceList = "/admin.service.v1.InternalMessageCategoryService/List"
const OperationInternalMessageCategoryServiceUpdate = "/admin.service.v1.InternalMessageCategoryService/Update"

type InternalMessageCategoryServiceHTTPServer interface {
	// Create 创建站内信消息分类
	Create(context.Context, *v11.CreateInternalMessageCategoryRequest) (*emptypb.Empty, error)
	// Delete 删除站内信消息分类
	Delete(context.Context, *v11.DeleteInternalMessageCategoryRequest) (*emptypb.Empty, error)
	// Get 查询站内信消息分类详情
	Get(context.Context, *v11.GetInternalMessageCategoryRequest) (*v11.InternalMessageCategory, error)
	// List 查询站内信消息分类列表
	List(context.Context, *v1.PagingRequest) (*v11.ListInternalMessageCategoryResponse, error)
	// Update 更新站内信消息分类
	Update(context.Context, *v11.UpdateInternalMessageCategoryRequest) (*emptypb.Empty, error)
}

func RegisterInternalMessageCategoryServiceHTTPServer(s *http.Server, srv InternalMessageCategoryServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/internal-message/categories", _InternalMessageCategoryService_List2_HTTP_Handler(srv))
	r.GET("/admin/v1/internal-message/categories/{id}", _InternalMessageCategoryService_Get2_HTTP_Handler(srv))
	r.POST("/admin/v1/internal-message/categories", _InternalMessageCategoryService_Create2_HTTP_Handler(srv))
	r.PUT("/admin/v1/internal-message/categories/{id}", _InternalMessageCategoryService_Update2_HTTP_Handler(srv))
	r.DELETE("/admin/v1/internal-message/categories/{id}", _InternalMessageCategoryService_Delete2_HTTP_Handler(srv))
}

func _InternalMessageCategoryService_List2_HTTP_Handler(srv InternalMessageCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationInternalMessageCategoryServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListInternalMessageCategoryResponse)
		return ctx.Result(200, reply)
	}
}

func _InternalMessageCategoryService_Get2_HTTP_Handler(srv InternalMessageCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetInternalMessageCategoryRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationInternalMessageCategoryServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetInternalMessageCategoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.InternalMessageCategory)
		return ctx.Result(200, reply)
	}
}

func _InternalMessageCategoryService_Create2_HTTP_Handler(srv InternalMessageCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateInternalMessageCategoryRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationInternalMessageCategoryServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreateInternalMessageCategoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _InternalMessageCategoryService_Update2_HTTP_Handler(srv InternalMessageCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateInternalMessageCategoryRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationInternalMessageCategoryServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateInternalMessageCategoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _InternalMessageCategoryService_Delete2_HTTP_Handler(srv InternalMessageCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteInternalMessageCategoryRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationInternalMessageCategoryServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteInternalMessageCategoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type InternalMessageCategoryServiceHTTPClient interface {
	// Create 创建站内信消息分类
	Create(ctx context.Context, req *v11.CreateInternalMessageCategoryRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Delete 删除站内信消息分类
	Delete(ctx context.Context, req *v11.DeleteInternalMessageCategoryRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Get 查询站内信消息分类详情
	Get(ctx context.Context, req *v11.GetInternalMessageCategoryRequest, opts ...http.CallOption) (rsp *v11.InternalMessageCategory, err error)
	// List 查询站内信消息分类列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListInternalMessageCategoryResponse, err error)
	// Update 更新站内信消息分类
	Update(ctx context.Context, req *v11.UpdateInternalMessageCategoryRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type InternalMessageCategoryServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewInternalMessageCategoryServiceHTTPClient(client *http.Client) InternalMessageCategoryServiceHTTPClient {
	return &InternalMessageCategoryServiceHTTPClientImpl{client}
}

// Create 创建站内信消息分类
func (c *InternalMessageCategoryServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreateInternalMessageCategoryRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/internal-message/categories"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationInternalMessageCategoryServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Delete 删除站内信消息分类
func (c *InternalMessageCategoryServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteInternalMessageCategoryRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/internal-message/categories/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationInternalMessageCategoryServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Get 查询站内信消息分类详情
func (c *InternalMessageCategoryServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetInternalMessageCategoryRequest, opts ...http.CallOption) (*v11.InternalMessageCategory, error) {
	var out v11.InternalMessageCategory
	pattern := "/admin/v1/internal-message/categories/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationInternalMessageCategoryServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 查询站内信消息分类列表
func (c *InternalMessageCategoryServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListInternalMessageCategoryResponse, error) {
	var out v11.ListInternalMessageCategoryResponse
	pattern := "/admin/v1/internal-message/categories"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationInternalMessageCategoryServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新站内信消息分类
func (c *InternalMessageCategoryServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateInternalMessageCategoryRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/internal-message/categories/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationInternalMessageCategoryServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}