# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Admin API
    description: GoWind Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/internal-message/categories:
        get:
            tags:
                - InternalMessageCategoryService
            description: 查询站内信消息分类列表
            operationId: InternalMessageCategoryService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: orderBy
                  in: query
                  description: 排序条件，其语法为JSON字符串，例如：{"val1", "-val2"}。字段名前加'-'为降序，否则为升序。
                  schema:
                    type: array
                    items:
                        type: string
                - name: query
                  in: query
                  description: AND过滤参数，其语法为json格式的字符串，如：{"key1":"val1","key2":"val2"}，具体请参见：https://github.com/tx7do/go-utils/tree/main/entgo/query/README.md
                  schema:
                    type: string
                - name: or
                  in: query
                  description: OR过滤参数，语法同AND过滤参数。
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListInternalMessageCategoryResponse'
        post:
            tags:
                - InternalMessageCategoryService
            description: 创建站内信消息分类
            operationId: InternalMessageCategoryService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateInternalMessageCategoryRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/internal-message/categories/{id}:
        get:
            tags:
                - InternalMessageCategoryService
            description: 查询站内信消息分类详情
            operationId: InternalMessageCategoryService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/InternalMessageCategory'
        put:
            tags:
                - InternalMessageCategoryService
            description: 更新站内信消息分类
            operationId: InternalMessageCategoryService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateInternalMessageCategoryRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - InternalMessageCategoryService
            description: 删除站内信消息分类
            operationId: InternalMessageCategoryService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
    /admin/v1/internal-message/messages:
        get:
            tags:
                - InternalMessageService
            description: 查询站内信消息列表
            operationId: InternalMessageService_ListMessage
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: orderBy
                  in: query
                  description: 排序条件，其语法为JSON字符串，例如：{"val1", "-val2"}。字段名前加'-'为降序，否则为升序。
                  schema:
                    type: array
                    items:
                        type: string
                - name: query
                  in: query
                  description: AND过滤参数，其语法为json格式的字符串，如：{"key1":"val1","key2":"val2"}，具体请参见：https://github.com/tx7do/go-utils/tree/main/entgo/query/README.md
                  schema:
                    type: string
                - name: or
                  in: query
                  description: OR过滤参数，语法同AND过滤参数。
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListInternalMessageResponse'
    /admin/v1/internal-message/messages/{id}:
        get:
            tags:
                - InternalMessageService
            description: 查询站内信消息详情
            operationId: InternalMessageService_GetMessage
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/InternalMessage'
        put:
            tags:
                - InternalMessageService
            description: 更新站内信消息
            operationId: InternalMessageService_UpdateMessage
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateInternalMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - InternalMessageService
            description: 删除站内信消息
            operationId: InternalMessageService_DeleteMessage
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreateInternalMessageCategoryRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/InternalMessageCategory'
            description: 创建站内信消息分类 - 请求
...
        InternalMessage:
            type: object
            properties:
                id:
                    type: integer
                    description: 消息ID
                    format: uint32
                title:
                    type: string
                    description: 消息标题
                content:
                    type: string
                    description: 消息内容
                status:
                    enum:
                        - DRAFT
                        - PUBLISHED
                        - SCHEDULED
                        - REVOKED
                        - ARCHIVED
                        - DELETED
                    type: string
                    description: 消息状态
                    format: enum
                type:
                    enum:
                        - NOTIFICATION
                        - PRIVATE
                        - GROUP
                    type: string
                    description: 消息类型
                    format: enum
                senderId:
                    type: integer
                    description: 发送者 ID（用户 ID，系统消息为0）
                    format: uint32
                senderName:
                    type: string
                    description: 发送者名称
                categoryId:
                    type: integer
                    description: 分类ID
                    format: uint32
                categoryName:
                    type: string
                    description: 分类名称
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 站内信消息
        InternalMessageCategory:
            type: object
            properties:
                id:
                    type: integer
                    description: 分类ID
                    format: uint32
                name:
                    type: string
                    description: 名称
                code:
                    type: string
                    description: 编码
                iconUrl:
                    type: string
                    description: 图标URL
                sortOrder:
                    type: integer
                    description: 排序顺序，值越小越靠前
                    format: uint32
                isEnabled:
                    type: boolean
                    description: 是否启用
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 站内信消息分类
...
        ListInternalMessageCategoryResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/InternalMessageCategory'
                total:
                    type: string
            description: 查询站内信消息分类列表 - 回应
        ListInternalMessageResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/InternalMessage'
                total:
                    type: string
            description: 查询站内信消息列表 - 回应
...
        UpdateInternalMessageCategoryRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/InternalMessageCategory'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新站内信消息分类 - 请求
        UpdateInternalMessageRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/InternalMessage'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新站内信消息 - 请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...